# go/openapi-generator/gen/go/model_order_request.go
/*
 * Sample API
 *
 * API documentation example using OpenAPI 3.0
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type OrderRequest struct {

	ProjectId int32 `json:"project_id,omitempty"`

	ReturnId int32 `json:"return_id,omitempty"`
}